+++ C/multi_agents/mpi_utils.c
@@ -27,7 +27,7 @@ int mpi_init() {
   return rank;
 }
 
-int mpi_perform_update(model* m, double lr, int scale) {
+int mpi_perform_update(model* m, float lr, int scale) {
   if (DEBUG_MPI) {
     printf("%d Entering mpi perform update\n", rank);
   }
@@ -36,9 +36,9 @@ int mpi_perform_update(model* m, double lr, int scale) {
   }
   matrix_t* l_grads = flatten_grad(m);
   matrix_t* g_grads = new_matrix(l_grads->rows, l_grads->cols);
-  MPI_Allreduce(l_grads->data, g_grads->data, l_grads->rows*l_grads->cols, MPI_DOUBLE, MPI_SUM, MPI_COMM_WORLD);
+  MPI_Allreduce(l_grads->data, g_grads->data, l_grads->rows*l_grads->cols, MPI_FLOAT, MPI_SUM, MPI_COMM_WORLD);
   if (scale) {
-    mult_scalar(g_grads, 1/(double)world_size);
+    mult_scalar(g_grads, 1/(float)world_size);
   }
   set_grad(m, g_grads);
   perform_update(m, lr);
@@ -65,12 +65,12 @@ int mpi_update_normalizer(normalizer* norm, matrix_t** data, int count) {
   matrix_t* g_avg_sumsq = new_matrix(l_sumsq->rows, l_sumsq->cols);
 
   MPI_Allreduce(&l_n, &g_n, 1, MPI_INT, MPI_SUM, MPI_COMM_WORLD);
-  MPI_Allreduce(l_sum->data, g_avg_sum->data, l_sum->rows*l_sum->cols, MPI_DOUBLE, MPI_SUM, MPI_COMM_WORLD);
-  MPI_Allreduce(l_sumsq->data, g_avg_sumsq->data, l_sumsq->rows*l_sumsq->cols, MPI_DOUBLE, MPI_SUM, MPI_COMM_WORLD);
+  MPI_Allreduce(l_sum->data, g_avg_sum->data, l_sum->rows*l_sum->cols, MPI_FLOAT, MPI_SUM, MPI_COMM_WORLD);
+  MPI_Allreduce(l_sumsq->data, g_avg_sumsq->data, l_sumsq->rows*l_sumsq->cols, MPI_FLOAT, MPI_SUM, MPI_COMM_WORLD);
 
   g_n /= world_size;
-  mult_scalar(g_avg_sum, 1/(double)world_size);
-  mult_scalar(g_avg_sumsq, 1/(double)world_size);
+  mult_scalar(g_avg_sum, 1/(float)world_size);
+  mult_scalar(g_avg_sumsq, 1/(float)world_size);
 
   free_matrix(norm->sum);
   free_matrix(norm->sumsq);
@@ -94,7 +94,7 @@ int mpi_sync(model* m) {
     printf("%d Entering mpi sync\n", rank);
   }
   matrix_t* w = flatten_weights(m);
-  MPI_Bcast(w->data, w->cols*w->rows, MPI_DOUBLE, 0, MPI_COMM_WORLD);
+  MPI_Bcast(w->data, w->cols*w->rows, MPI_FLOAT, 0, MPI_COMM_WORLD);
   MPI_Barrier(MPI_COMM_WORLD);
   set_weights(m, w);
   free_matrix(w);
@@ -110,12 +110,12 @@ int mpi_check_sync(model* m) {
   }
   if (!rank) {
     matrix_t* w = flatten_weights(m);
-    MPI_Bcast(w->data, w->cols*w->rows, MPI_DOUBLE, 0, MPI_COMM_WORLD);
+    MPI_Bcast(w->data, w->cols*w->rows, MPI_FLOAT, 0, MPI_COMM_WORLD);
     free_matrix(w);
   } else {
     matrix_t* l_w = flatten_weights(m);
     matrix_t* root_w = new_matrix(l_w->rows, l_w->cols);
-    MPI_Bcast(root_w->data, l_w->rows*l_w->cols, MPI_DOUBLE, 0, MPI_COMM_WORLD);
+    MPI_Bcast(root_w->data, l_w->rows*l_w->cols, MPI_FLOAT, 0, MPI_COMM_WORLD);
     assert(equal(l_w, root_w));
     free_matrix(root_w);
     free_matrix(l_w);
